add modules to project
