+++ src/index.html
@@ -21,7 +21,7 @@
         <i class="fa-solid fa-arrow-turn-down fa-rotate-90"></i>
       </div>
       <div>
-        <ul class="to-do-list-container"></ul>
+        <ul class="list-container"></ul>
       </div>
       <button id="clear-btn">Clear all completed</button>
     </section>
