+++ src/index.js
@@ -1,29 +1,7 @@
 import './style.css';
+import itemsListDisplay from './modules/listItems.js';
 
-const itemsList = [{
-  index: 0,
-  description: 'Finish the to-do list',
-  completed: false,
-},
-{
-  index: 1,
-  description: 'Clean the house',
-  completed: false,
-},
-{
-  index: 2,
-  description: 'wash the dishes',
-  completed: true,
-},
-];
-
-const toDoListContainer = document.querySelector('.to-do-list-container');
-
-const itemsListDisplay = () => {
-  itemsList.forEach((item) => {
-    toDoListContainer.innerHTML += `
-    <li class='to-do-item' value='index'>${item.description}<i class="fa-solid fa-ellipsis-vertical"></i></li>`;
-  });
+const render = () => {
+  window.addEventListener('load', itemsListDisplay());
 };
-
-window.addEventListener('load', itemsListDisplay);
+render();
