+++ src/modules/listData.js
@@ -1,0 +1,18 @@
+const itemsList = [{
+  index: 0,
+  description: 'Finish the to-do list',
+  completed: false,
+},
+{
+  index: 1,
+  description: 'Clean the house',
+  completed: false,
+},
+{
+  index: 2,
+  description: 'wash the dishes',
+  completed: true,
+},
+];
+
+export default itemsList;
